Fix package search to handle @types packages
Typescript now installs type declarations in subdirectories under
node_modules/@types. This adds code to correctly include these
packages when building the list as well as when printing the
packages titles in the report.

Also adds a small change to work on Node on Windows. The way node
recursively call node.cli results in a couple of extra arguments in
opt._ that we need to remove. This may be a bug in minimist but
seems pretty easy to work around.